+++ doc/changelog.md
@@ -32,7 +32,7 @@ Syntax: `- short text describing the change _(Your Name)_`
 - Add toggleable plant labels _(Moritz)_
 - _()_
 - _()_
-- _()_
+- Fix various validation bugs in seed form _(Moritz)_
 - _()_
 - Link seeds page in the navbar _(Moritz)_
 - _()_
@@ -42,7 +42,7 @@ Syntax: `- short text describing the change _(Your Name)_`
 - _()_
 - Enable seeds to be edited and deleted _(Giancarlo & Moritz)_
 - Update UC for timeline and plants layer _(Paul)_
-- _()_
+- Fix bug in updating of the plant selection _(Paul)_
 - _()_
 - _()_
 - Added Meeting Agenda for 28.8. 9:00 _(Markus Raab)_
@@ -60,14 +60,23 @@ Syntax: `- short text describing the change _(Your Name)_`
 - _()_
 - _()_
 - Add Christoph Nemeth as Teammember _(tschawax)_
-- Fix middle mouse button to only move the stage or select individual plants _(tschawax)_
+- Fix middle mouse button to only move the stage or select individual plants #817 _(tschawax)_
+- Fix selections being kept after choosing a plant from the search list #787 _(tschawax)_
+- Fix placing new plants onto existing plants #890 _(tschawax)_
+- Fix select box while being in placement mode #886 _(tschawax)_
 - _()_
+- _()_
+- Seeds: translate quality and quantity columns _(Moritz)_
 - _()_
 - Fix drag select not working and add appropriate e2e regression tests #852 _(4ydan & Moritz)_
 - _()_
 - _()_
 - _()_
 - _()_
+- _()_
+- CI: Update doc _(4ydan)_
+- Doc: Update testing strategy #864 _(4ydan)_
+- E2E: Add new e2e test for undoing deletions #889 _(4ydan)_
 - CI: Add changelog test to CI, only in PR stages #857 _(4ydan)_
 - E2E: Worker uuids fixture #837 _(4ydan)_
 - GIT: Add data analysis jupyter notebook #843 _(4ydan)_